Adding Email Function

+++ app/controllers/backoffice/admins_controller.rb
@@ -1,6 +1,6 @@
 class Backoffice::AdminsController < BackofficeController
   before_action :set_admin, only: [:edit, :update, :destroy]
-  after_action :verify_authorized, only: :new
+  after_action :verify_authorized, only: [:new, :destroy]
   after_action :verify_policy_scoped, only: :index
   
   def index
@@ -29,6 +29,7 @@ class Backoffice::AdminsController < BackofficeController
   
   def update
     if @admin.update(params_admin)
+      AdminMailer.update_email(current_admin, @admin).deliver_now
       redirect_to backoffice_admins_path, notice: "Update Successful."
     else
       render :edit
@@ -37,6 +38,7 @@ class Backoffice::AdminsController < BackofficeController
   end
 
   def destroy
+    authorize @admin
     if @admin.destroy
       redirect_to backoffice_admins_path, notice: "Delete Successful."
     else
